Check for defaultPrevented
Added check for defaultPrevented, because you probably don't want to call the callback (for example, doing a push-state) if the default was already prevented.  Alternatively, you could return the event object so that it the callback could make this determination for itself (ie. `cb(url.resolve(location.href, href), ev);`